Sign up is work

+++ src/main/java/ru/kivit/services/UserService.java
@@ -1,21 +1,23 @@
 package ru.kivit.services;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import ru.kivit.models.Role;
 import ru.kivit.models.User;
 import ru.kivit.repositories.UserRepository;
 
 import java.util.List;
 
 @Service
-public class UserService {
+@AllArgsConstructor
+public class UserService implements UserDetailsService {
 
     private final UserRepository repository;
-
-    @Autowired
-    public UserService(UserRepository repository) {
-        this.repository = repository;
-    }
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     public User findById(Long id){
         return repository.getOne(id);
@@ -25,8 +27,20 @@ public class UserService {
         return repository.findAll();
     }
 
-    public User save(User user){
-        return repository.save(user);
+    public boolean save(User user){
+
+        User userFromDb = repository.findByLogin(user.getLogin());
+
+        if (userFromDb != null){
+            return false;
+        }
+
+        final String encryptedPassword = bCryptPasswordEncoder.encode(user.getPassword());
+        user.setPassword(encryptedPassword);
+        user.setRole(Role.USER);
+        repository.save(user);
+
+        return true;
     }
 
     public User update(User newUser, Long id){
@@ -34,11 +48,25 @@ public class UserService {
         oldUser.setFirst_name(newUser.getFirst_name());
         oldUser.setLast_name(newUser.getLast_name());
         oldUser.setLogin(newUser.getLogin());
-        oldUser.setPassword(newUser.getPassword());
+        oldUser.setPassword(bCryptPasswordEncoder.encode(newUser.getPassword()));
         return  repository.save(oldUser);
     }
 
-    public void deleteById(Long id){
-        repository.deleteById(id);
+    public boolean deleteById(Long id){
+
+        if (repository.findById(id).isPresent()) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String login) throws UsernameNotFoundException {
+        User userFromDb = repository.findByLogin(login);
+        if (userFromDb == null){
+            throw new UsernameNotFoundException("Incorrect login");
+        }
+        return userFromDb;
     }
 }
